+++ app/lib/tool/backfill/backfill_new_fields.dart
@@ -6,8 +6,6 @@ import 'package:clock/clock.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:pub_dev/account/models.dart';
-import 'package:pub_dev/package/api_export/api_exporter.dart';
-import 'package:pub_dev/package/backend.dart';
 import 'package:pub_dev/package/models.dart';
 import 'package:pub_dev/publisher/models.dart';
 import 'package:pub_dev/shared/datastore.dart';
@@ -28,34 +26,6 @@ Future<void> backfillNewFields() async {
 @visibleForTesting
 Future<void> migrateIsBlocked() async {
   _logger.info('Migrating isBlocked...');
-  final pkgQuery = dbService.query<Package>()..filter('isBlocked =', true);
-  await for (final entity in pkgQuery.run()) {
-    await withRetryTransaction(dbService, (tx) async {
-      final pkg = await tx.lookupValue<Package>(entity.key);
-      // sanity check
-      if (!pkg.isBlocked) {
-        return;
-      }
-      pkg
-        ..isModerated = true
-        ..moderatedAt = pkg.moderatedAt ?? pkg.blocked ?? clock.now()
-        ..isBlocked = false
-        ..blocked = null
-        ..blockedReason = null;
-      tx.insert(pkg);
-    });
-
-    // sync exported API(s)
-    await apiExporter?.synchronizePackage(entity.name!, forceDelete: true);
-
-    // retract or re-populate public archive files
-    await packageBackend.tarballStorage.updatePublicArchiveBucket(
-      package: entity.name!,
-      ageCheckThreshold: Duration.zero,
-      deleteIfOlder: Duration.zero,
-    );
-  }
-
   final publisherQuery = dbService.query<Publisher>()
     ..filter('isBlocked =', true);
   await for (final entity in publisherQuery.run()) {
@@ -95,12 +65,12 @@ Future<void> migrateIsBlocked() async {
 Future<void> _removeKnownUnmappedFields() async {
   await for (final p in dbService.query<Package>().run()) {
     if (p.additionalProperties.isEmpty) continue;
-    if (p.additionalProperties.containsKey('isWithheld') ||
-        p.additionalProperties.containsKey('withheldReason')) {
+    if (p.additionalProperties.containsKey('blocked') ||
+        p.additionalProperties.containsKey('blockedReason')) {
       await withRetryTransaction(dbService, (tx) async {
         final pkg = await tx.lookupValue<Package>(p.key);
-        pkg.additionalProperties.remove('isWithheld');
-        pkg.additionalProperties.remove('withheldReason');
+        pkg.additionalProperties.remove('blocked');
+        pkg.additionalProperties.remove('blockedReason');
         tx.insert(pkg);
       });
     }